requirements.txt
absl-py==2.1.0
astunparse==1.6.3
attrs==23.2.0
cachetools==5.4.0
certifi==2024.7.4
cffi==1.16.0
charset-normalizer==3.3.2
contourpy==1.2.1
cycler==0.12.1
exceptiongroup==1.2.2
flatbuffers==24.3.25
fonttools==4.53.1
gast==0.4.0
google-auth==2.32.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.65.2
h11==0.14.0
h5py==3.11.0
idna==3.7
importlib_metadata==8.2.0
importlib_resources==6.4.0
joblib==1.4.2
keras==2.10.0
Keras-Preprocessing==1.1.2
kiwisolver==1.4.5
libclang==18.1.1
Markdown==3.6
MarkupSafe==2.1.5
matplotlib==3.9.1
mysql-connector-python==9.0.0
numpy==1.26.4
oauthlib==3.2.2
opt-einsum==3.3.0
outcome==1.3.0.post0
packaging==24.1
pandas==2.2.2
pillow==10.4.0
protobuf==3.19.6
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
pyparsing==3.1.2
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
schedule==1.2.2
scikit-learn==1.5.1
scipy==1.13.1
seaborn==0.13.2
selenium==4.23.1
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
tensorboard==2.10.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.10.1
tensorflow-estimator==2.10.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.4.0
threadpoolctl==3.5.0
trio==0.26.0
trio-websocket==0.11.1
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
webdriver-manager==4.0.2
websocket-client==1.8.0
Werkzeug==3.0.3
wrapt==1.16.0
wsproto==1.2.0
zipp==3.19.2
